Too many changes :(
Sadly bit too many changes. But basically I have fixed the problems with
ball creation that existed. I have added a tennisball class, I have
added the assets to the repo but i think I should probably remove them
and try to change the location of the repo to be further up and just rm
the dirs that I dont want to use.  I added a try/catch for the
unexpected exits that would occur at the loading screen, these seem to
be related to the androidScreen class and I cannot really handle that.
Now i need to playtest some more, adjust various constants, add proper
resources, add a life indicator and some other things in the todo file.