Collisions declared fixed
Very sporadically I can still find issues with colissions but I am not
able to reproduce them reliably. Added some new asset. I think I can
call it a pretty productive day, if I focus on it I think I can pulish
tonight, just need to tweak values. For future reference, the changes
that I made were to first make sure that if a drag event has no locale
change between first and last, that it is considered a click instead of
a drag. Second I changed the way drag events calculate force to not use
a fake ball but to simply use the relative positions of the click event
and the ball and multiply it by a factor depending on the percieved drag
speed.