Tweaks and cleanup
Tweaking the numbers used in the game, the weight of the ball, the
strength of gravity etc to make it more fun. Also cleaned up the code a
bit, removing commented sections with posX and posY instead of a pos
vector, removed some other old commented code snippets and unused
variables. Also fixed the way tennisBall is handled and the
createTEnnisBall function should now work as intended. Also cleaned up a
bit in the update function, making a separate checkEdges function to
handle bouncing with walls and within that function implemented min/max
y/x positions instead of manually writing those in the places they are
used to make it easier to for example add an area of the screen not in
used by the game.