Added features for tennisballs and multiple sounds
Added tennisball bouncing in inBall . Also added support for playing one
randomly selected out of multiple sounds for kicks and bounces (though
there is currently only one bounce sound). Oh and added playing sounds I
guess. Also changed the checkEdge method to take a ball instead of an
int so that it can check if it's a tennisball or a football and if the
former remove it, and if the latter die etc.