Adding tennisball
Currently in the process of adding tennisball. Currently there is no
purpose other than being another thing, and not yet tested if works but
added drawing, colission and destruction of a single tennisball. Can
relatively easily be modified to be a list. In  fact tha gives me an
idea to change the inBall feature to return a ball which could be null
if nothing is found?