Fixed a bug where balls collided with themselves
The recent changes in the updateBalls method changed a double for loops
over all balls to a single for loop over all balls with a call to
inBall. However this did not check if the collided with target was the
same ball resulting in a colission with self and an inf/NaN force since
the distance between them is zero.