Some nice improvements
Added a nicer lives left indicator with small balls, added a function to
draw a scaled image at a certain position . Added a delay to the
endscreen so that one cannot continue to a new game immediately (and
thus miss the grade). Added a few more reviews and fixed a couple of
spelling errors in the reviews as well as removing one non-PC review.
Added some comments to the gameScreen class.