Large refactoring and added spin
Moved all functions checking and updating the balls and game state to a
new class called Engine. Added a rotating feature to android graphics
and added a spin and angle variable to the Ball class. Added spin when
colliding with walls and when clicking, not yet added to dragging (since
that isn't working correctly) nor to ball-ball collisions.