Making ammends
Have not been uploading to github. Tested out various stuff with
different time steps.
Changed the way rotation is performed that should improve performance
noticeably.

Changed the way touches are handled to support holding a finger down.

Added a large part of the support for syncing a highscore to a remote
server, not yet implemented a server etc.

Also made changes to how spin is calculated

Currently working on making dragging working as intended with variable
strength depending on finger speed.