Finished changes to dragging/swiping
Now dragging has a minimum speed under which it registers as a click,
and a maximum speed which it is limited to. Should still implement spin
interaction in ball collisions. Also removed support for tablets (I
hope) but that is actually outside of the codebase covered in this repo,
but you know.